refactor: consolidate CookieAuth and improve JsonRpcConfig
- Move CookieAuth from zainod to zaino-commons for shared usage
- Update JsonRpcConfig to use unified CookieAuth enum instead of separate boolean + optional fields
- Rename json_rpc_listen_address to listen_address (less redundant)
- Add zaino-commons dependency to zaino-serve
- Update all usage sites in zainod, zaino-serve, and zaino-testutils
- Simplify cookie authentication setup using CookieAuth pattern matching

This creates a consistent approach across all server configs:
- CookieAuth::Enabled { path } vs CookieAuth::Disabled
- Reuses shared cookie reading logic from zaino-commons
- Cleaner, more maintainable configuration structure
