+++ zainod/src/indexer.rs
@@ -94,12 +94,8 @@ where
                 JsonRpcServer::spawn(
                     service.inner_ref().get_subscriber(),
                     JsonRpcConfig {
-                        json_rpc_listen_address: indexer_config.server.json_rpc_listen_address,
-                        enable_cookie_auth: matches!(indexer_config.server.cookie, crate::config::CookieAuth::Enabled { .. }),
-                        cookie_dir: match &indexer_config.server.cookie {
-                            crate::config::CookieAuth::Enabled { path } => Some(path.clone()),
-                            crate::config::CookieAuth::Disabled => None,
-                        },
+                        listen_address: indexer_config.server.json_rpc_listen_address,
+                        auth: indexer_config.server.cookie.clone(),
                     },
                 )
                 .await
